Added binBoxGen.py
This is a template program that takes L, W, H, t as input parameteres and outputs an SVG image of the bin box template for those dimensions. I set up the second half of the program to make a polygon with 5 random sets of points I put in. All we would need to do is change those points to make a different polygon